acceptance_tests/pyproject.toml
[tool.poetry]
name = "shared-config-manager-acceptance"
version = "0.0.0"
description = "Not used"
authors = []

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pytest-cov = "3.0.0"
prospector = "1.7.7"
typing-extensions = "4.3.0"
c2cwsgiutils = {version = "5.1.3", extras = ["standard", "test_images"]}
requests = "2.28.1"
boltons = "21.0.0"
netifaces = "0.11.0"
lxml = "4.9.1"
